requirements.txt
fastapi
uvicorn
httpx
pyyaml
python-dotenv
openai
